Adding arrays everywhere
This is a lot more than the two previous. I added arrays everywhere that I could.

The only place it didn't work like I thought it should is on the work.php page. The job descriptions would be long and I'm not sure what the best practices are to keep it readable.

Also, I'm imagining that each theatre will get their own public profile, so I created one of those and put it in a folder to keep stuff organized. But I think that was totally wrong since it absolutely doesn't work.